+++ mail_deduplicate/strategy.py
@@ -23,7 +23,7 @@ import random
 import re
 from functools import wraps
 
-from click_extra.colorize import default_theme as theme
+from click_extra import get_default_theme
 
 TYPE_CHECKING = False
 if TYPE_CHECKING:
@@ -34,6 +34,9 @@ if TYPE_CHECKING:
     from .mail import DedupMailMixin
 
     SelectionFunc: TypeAlias = Callable[[DuplicateSet], set[DedupMailMixin]]
+
+
+theme = get_default_theme()
 
 
 def log_selection(message_template: str):